Add utils/node/plugin-action
